+++ CalculatorApi/Program.cs
@@ -1,11 +1,10 @@
+using Lab_1;
+using Microsoft.AspNetCore.Mvc;
 var builder = WebApplication.CreateBuilder(args);
 
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
-builder.Services.AddSingleton<Add>();
-builder.Services.AddSingleton<Subtract>();
-builder.Services.AddSingleton<Multiply>();
-builder.Services.AddSingleton<Divide>();
+builder.Services.AddSingleton<Calculator>();
 
 var app = builder.Build();
 
@@ -17,25 +16,25 @@ if (app.Environment.IsDevelopment())
 }
 
 app.UseHttpsRedirection();
-app.MapGet("/add", (double a, double b, [FromServices] Add calculator) =>
+app.MapGet("/add", (double a, double b, [FromServices] Calculator calculator) =>
 {
     return calculator.Add(a, b);
 })
 .WithOpenApi();
 
-app.MapGet("/subtract", (double a, double b, [FromServices] Subtract calculator) =>
+app.MapGet("/subtract", (double a, double b, [FromServices] Calculator calculator) =>
 {
     return calculator.Subtract(a, b);
 })
 .WithOpenApi();
 
-app.MapGet("/divide", (double a, double b, [FromServices] Divide calculator) =>
+app.MapGet("/divide", (double a, double b, [FromServices] Calculator calculator) =>
 {
     return calculator.Divide(a, b);
 })
 .WithOpenApi();
 
-app.MapGet("/multiply", (double a, double b, [FromServices] Multiply calculator) =>
+app.MapGet("/multiply", (double a, double b, [FromServices] Calculator calculator) =>
 {
     return calculator.Multiply(a, b);
 })
