--- CalculatorLibrary/Add.cs
@@ -1,10 +1,0 @@
-namespace Lab_1
-{
-    public class Add : IAdd
-    {
-        public double Add(double a, double b)
-        {
-            return a + b;
-        }
-    }
-}
