--- CalculatorLibrary/Divide.cs
@@ -1,10 +1,0 @@
-namespace Lab_1
-{
-    public class Divide : IDivide
-    {
-        public double Divide(double a, double b)
-        {
-            return a / b;
-        }
-    }
-}
