--- CalculatorLibrary/Multiply.cs
@@ -1,10 +1,0 @@
-namespace Lab_1
-{
-	public class Multiply : IMultiply
-	{
-		public double Multiply(double a, double b)
-		{
-			return a * b;
-		}
-	}
-}
