--- CalculatorLibrary/Subtract.cs
@@ -1,10 +1,0 @@
-namespace Lab_1
-{
-    public class Subtract : ISubtract
-    {
-        public double Subtract(double a, double b)
-        {
-            return a - b;
-        }
-    }
-}
